refactor: split client into each file
