+++ api/incident.go
@@ -1,5 +1,7 @@
 package api
 
+import "encoding/json"
+
 type IncidentOptions struct {
 	Incident struct {
 		// common options
@@ -144,4 +146,29 @@ func newIncidentOptions(opts ...IncidentOption) *IncidentOptions {
 	}
 
 	return options
+}
+
+func (cli *Client) CreateIncident(opts ...IncidentOption) (*Response, error) {
+	incidentOpts := newIncidentOptions(opts...)
+	marshalled, err := json.Marshal(incidentOpts)
+	if err != nil {
+		return nil, err
+	}
+
+	return cli.Post(URL("/api/v1/user/incidents/"), &Request{
+		Raw: json.RawMessage(marshalled),
+	})
+}
+
+func (cli *Client) UpdateIncident(id string, opts ...IncidentOption) (*Response, error) {
+	incidentOpts := newIncidentOptions(opts...)
+	marshalled, err := json.Marshal(incidentOpts)
+	if err != nil {
+		return nil, err
+	}
+
+	url := URL("/api/v1/user/incidents/%s/", id)
+	return cli.Put(url, &Request{
+		Raw: json.RawMessage(marshalled),
+	})
 }
